move to packages, add maven, readme, gitignore
