+++ readme.md
@@ -1,0 +1,5 @@
+## Turing Emulator
+
+Process Turing Machine commands
+
+run with `src/test/resources/turing.json` argument for testing
