+++ src/main/java/ru/mephi22/turing/interfaces/InfiniteTape.java
@@ -1,0 +1,8 @@
+package ru.mephi22.turing.interfaces;
+
+public interface InfiniteTape {
+    char getInput();
+    void setOutput(char sym);
+    void goLeft();
+    void goRight();
+}
