--- src/tMachine/interfaces/InfiniteTape.java
@@ -1,8 +1,0 @@
-package tMachine.interfaces;
-
-public interface InfiniteTape {
-    public char getInput();
-    public void setOutput(char sym);
-    public void goLeft();
-    public void goRight();
-}
