--- src/tMachine/interfaces/RuleStore.java
@@ -1,9 +1,0 @@
-package tMachine.interfaces;
-
-import tMachine.RightRule;
-import tMachine.LeftRule;
-
-public interface RuleStore {
-    public void addRule( LeftRule cs, RightRule csm);
-    public RightRule getRule(String state, String sym);
-}
