+++ src/test/resources/turing.json
@@ -1,0 +1,16 @@
+{
+  "rule": [
+    {
+      "src":"S0{∂}",
+      "dst":"{∂}{R}S0",
+    },
+    {
+      "src":"S0{a}",
+      "dst":"{b}{R}S0"},
+    {
+      "src":"S0{λ}",
+      "dst":"{λ}{H}Ω",
+    }
+  ],
+  "str":["∂aabbbaaabb"],
+}
